wl18xx: pg2 update
pg2 is recognized with its unique chip id.
pg2 differ from pg1 in some control registers and addresses.
the difference in addressing in now maintained as a private struct
in the wl18xx_pric struct, that is assigned after recognizing the chip.
the yield workaround uses the specific struct to use the appropriate
addresses.
since pg2 suppose to work without the yield workaround, a new module
parameter is added, disable_yield_fix, that if assigned to true, the
workaround is discarded. default value is false.

Signed-off-by: Ido Reis <idor@ti.com>